Further Peer Comments from @Fryguy updated in VixDiskLib binding PR.
- Remove extraneous blank lines.
- Remove ");" in API comments copied from VixDiskLib.h file.
- Make methods private when expected.
- Remove unneeded comments (commented out code) & align other comments
- Simplify x.nil? ? nil : foo(x) construct to x && foo(x)
- Remove unnecessary server disk count hash (simple variable suffices)
- replace string concatenation with interpolation
- remove unnecessary init log arguments from client & server and mark
them as ignored by preceding the names with "_" character.
- rename confusing "binding" variable.
- Require VixDiskLibClient init to be called before connect but
do not preclude a second init call from occurring.
- Move individual classes into separate files.  This includes the
DiskInfo, SafeConnectParams, and SafeCreateParams classes from the
vixdisklib_api.rb file, and the VDDKFactory class with associated DRb
startup functionality has been separated from the original VixDiskLib
wrapper classes (VixDiskLib, VdlConnection, & VdlDisk)
